Fixing read .csv still (fixed)

+++ final_launch.py
@@ -247,8 +247,8 @@ class Infos:
             row = row[0]
             # print(row)
             row[1] = int(row[1])
-            row[2] = int(row[2])
-            row[3] = int(row[3])
+            row[2] = float(row[2])
+            row[3] = float(row[3])
             # row = readCSV[time_idx]
             if self.prev_turbulences:
                 if len(self.prev_turbulences) == 5: # TODO could edit for more effects
